fixes regression from PR: 'Typed array cleanup' (issue #1919)

+++ tests/testsrc/jstests/es6/bigint64array.js
@@ -5,22 +5,22 @@
 load("testsrc/assert.js");
 
 (function TestSymbolSpecies() {
-  var symbolSpeciesValue = Uint8ClampedArray[Symbol.species];
-  assertEquals(Uint8ClampedArray, symbolSpeciesValue);
+  var symbolSpeciesValue = BigInt64Array[Symbol.species];
+  assertEquals(BigInt64Array, symbolSpeciesValue);
 })();
 
 (function TestSymbolToString() {
-  var a = new Uint8ClampedArray(4);
-  assertEquals('Uint8ClampedArray', a[Symbol.toStringTag]);
+  var a = new BigInt64Array(4);
+  assertEquals('BigInt64Array', a[Symbol.toStringTag]);
   assertEquals(false, a.hasOwnProperty(Symbol.toStringTag));
 
-  assertEquals(undefined, Uint8ClampedArray[Symbol.toStringTag]);
-  assertEquals(false, Uint8ClampedArray.hasOwnProperty(Symbol.toStringTag));
-  assertEquals(false, Uint8ClampedArray.prototype.hasOwnProperty(Symbol.toStringTag));
+  assertEquals(undefined, BigInt64Array[Symbol.toStringTag]);
+  assertEquals(false, BigInt64Array.hasOwnProperty(Symbol.toStringTag));
+  assertEquals(false, BigInt64Array.prototype.hasOwnProperty(Symbol.toStringTag));
 })();
 
 (function TestPrototypeSymbolToString() {
-  var a = new Uint8ClampedArray(4).__proto__;
+  var a = new BigInt64Array(4).__proto__;
   assertEquals(undefined, a[Symbol.toStringTag]);
   assertEquals(false, a.hasOwnProperty(Symbol.toStringTag));
 })();
